Move SmartScalar subcomponents to separate files
