+++ frontend/src/metabase/visualizations/visualizations/SmartScalar/SmartScalar.tsx
@@ -1,18 +1,11 @@
 import cx from "classnames";
-import { type PropsWithChildren, useEffect, useMemo, useRef } from "react";
-import innerText from "react-innertext";
-import { jt, t } from "ttag";
+import { useEffect, useMemo, useRef } from "react";
+import { t } from "ttag";
 
-import { Ellipsified } from "metabase/common/components/Ellipsified";
 import DashboardS from "metabase/css/dashboard.module.css";
-import { getIsNightMode } from "metabase/dashboard/selectors";
-import { color, lighten } from "metabase/lib/colors";
-import { formatValue } from "metabase/lib/formatting/value";
-import { measureTextWidth } from "metabase/lib/measure-text";
-import { useSelector } from "metabase/lib/redux";
-import { isEmpty } from "metabase/lib/validate";
+import { color } from "metabase/lib/colors";
 import EmbedFrameS from "metabase/public/components/EmbedFrame/EmbedFrame.module.css";
-import { Box, Flex, Text, Title, Tooltip, useMantineTheme } from "metabase/ui";
+import { Box } from "metabase/ui";
 import {
   ScalarValue,
   ScalarWrapper,
@@ -26,7 +19,6 @@ import {
   getMinSize,
 } from "metabase/visualizations/shared/utils/sizes";
 import type {
-  ColumnSettings,
   VisualizationDefinition,
   VisualizationPassThroughProps,
   VisualizationProps,
@@ -34,25 +26,16 @@ import type {
 
 import { ScalarContainer } from "../Scalar/Scalar.styled";
 
+import { PreviousValueComparison } from "./PreviousValueComparison";
+import { ScalarPeriod } from "./ScalarPeriod";
 import { SmartScalarComparisonWidget } from "./SettingsComponents/SmartScalarSettingsWidgets";
-import { VariationIcon, VariationValue } from "./SmartScalar.styled";
-import {
-  CHANGE_TYPE_OPTIONS,
-  type ComparisonResult,
-  computeTrend,
-} from "./compute";
+import { computeTrend } from "./compute";
 import {
   DASHCARD_HEADER_HEIGHT,
-  ICON_MARGIN_RIGHT,
-  ICON_SIZE,
   MAX_COMPARISONS,
-  SPACING,
-  TOOLTIP_ICON_SIZE,
   VIZ_SETTINGS_DEFAULTS,
 } from "./constants";
 import {
-  formatChangeAutoPrecision,
-  getChangeWidth,
   getColumnsForComparison,
   getComparisonOptions,
   getComparisons,
@@ -157,236 +140,6 @@ export function SmartScalar({
         </Box>
       ))}
     </ScalarWrapper>
-  );
-}
-
-interface ScalarPeriodProps {
-  period: string | number | JSX.Element | null;
-  onClick?: () => void;
-}
-
-function ScalarPeriod({ period, onClick }: ScalarPeriodProps) {
-  return (
-    <Text
-      data-testid="scalar-period"
-      component="h3"
-      ta="center"
-      style={{ cursor: onClick && "pointer" }}
-      fw={700}
-      lh="1rem"
-      className={cx(
-        DashboardS.fullscreenNormalText,
-        DashboardS.fullscreenNightText,
-        EmbedFrameS.fullscreenNightText,
-      )}
-      onClick={onClick}
-    >
-      <Ellipsified tooltip={period} lines={1} placement="bottom">
-        {period}
-      </Ellipsified>
-    </Text>
-  );
-}
-
-const Separator = ({ inTooltip }: { inTooltip?: boolean }) => {
-  const theme = useMantineTheme();
-  const isNightMode = useSelector(getIsNightMode);
-
-  const separatorColor =
-    isNightMode || inTooltip
-      ? lighten(theme.fn.themeColor("text-medium"), 0.15)
-      : lighten(theme.fn.themeColor("text-light"), 0.25);
-
-  return (
-    <Text
-      mx="0.2rem"
-      style={{ transform: "scale(0.7)" }}
-      c={separatorColor}
-      component="span"
-    >
-      {" • "}
-    </Text>
-  );
-};
-
-interface PreviousValueComparisonProps {
-  comparison: ComparisonResult;
-  width: number;
-  fontFamily: string;
-  formatOptions: ColumnSettings;
-}
-
-function PreviousValueComparison({
-  comparison,
-  width,
-  fontFamily,
-  formatOptions,
-}: PreviousValueComparisonProps) {
-  const fontSize = "0.875rem";
-
-  const {
-    changeType,
-    percentChange,
-    comparisonDescStr,
-    comparisonValue,
-    changeArrowIconName,
-    changeColor,
-    display,
-  } = comparison;
-
-  const theme = useMantineTheme();
-  const isNightMode = useSelector(getIsNightMode);
-
-  const fittedChangeDisplay =
-    changeType === CHANGE_TYPE_OPTIONS.CHANGED.CHANGE_TYPE
-      ? formatChangeAutoPrecision(percentChange as number, {
-          fontFamily,
-          fontWeight: 900,
-          width: getChangeWidth(width),
-        })
-      : display.percentChange;
-
-  const availableComparisonWidth =
-    width -
-    4 * SPACING -
-    ICON_SIZE -
-    ICON_MARGIN_RIGHT -
-    measureTextWidth(innerText(<Separator />), {
-      size: fontSize,
-      family: fontFamily,
-      weight: 700,
-    }) -
-    measureTextWidth(fittedChangeDisplay, {
-      size: fontSize,
-      family: fontFamily,
-      weight: 900,
-    });
-
-  const valueCandidates = [
-    display.comparisonValue,
-    ...(changeType === CHANGE_TYPE_OPTIONS.CHANGED.CHANGE_TYPE
-      ? [formatValue(comparisonValue, { ...formatOptions, compact: true })]
-      : []),
-    "",
-  ];
-
-  const getDetailCandidate = (
-    valueFormatted: string | number | JSX.Element | null,
-    { inTooltip }: { inTooltip?: boolean } = {},
-  ) => {
-    if (isEmpty(valueFormatted)) {
-      return comparisonDescStr;
-    }
-
-    const descColor = inTooltip
-      ? "var(--mb-color-tooltip-text-secondary)"
-      : "var(--mb-color-text-secondary)";
-
-    if (isEmpty(comparisonDescStr)) {
-      return (
-        <Text key={valueFormatted as string} c={descColor} component="span">
-          {valueFormatted}
-        </Text>
-      );
-    }
-
-    return jt`${comparisonDescStr}: ${(
-      <Text key="value-str" c={descColor} component="span">
-        {valueFormatted}
-      </Text>
-    )}`;
-  };
-
-  const detailCandidates = valueCandidates.map((valueStr) =>
-    getDetailCandidate(valueStr),
-  );
-  const fullDetailDisplay = detailCandidates[0];
-  const fittedDetailDisplay = detailCandidates.find(
-    (e) =>
-      measureTextWidth(innerText(e), {
-        size: fontSize,
-        family: fontFamily,
-        weight: 700,
-      }) <= availableComparisonWidth,
-  );
-
-  const tooltipFullDetailDisplay = getDetailCandidate(valueCandidates[0], {
-    inTooltip: true,
-  });
-
-  const VariationPercent = ({
-    inTooltip,
-    iconSize,
-    children,
-  }: PropsWithChildren<{ inTooltip?: boolean; iconSize: string | number }>) => {
-    const noChangeColor =
-      inTooltip || isNightMode
-        ? lighten(theme.fn.themeColor("text-medium"), 0.3)
-        : "text-light";
-
-    return (
-      <Flex align="center" maw="100%" c={changeColor ?? noChangeColor}>
-        {changeArrowIconName && (
-          <VariationIcon name={changeArrowIconName} size={iconSize} />
-        )}
-        <VariationValue showTooltip={false}>{children}</VariationValue>
-      </Flex>
-    );
-  };
-
-  const VariationDetails = ({
-    inTooltip,
-    children,
-  }: PropsWithChildren<{ inTooltip?: boolean }>) => {
-    if (!children) {
-      return null;
-    }
-
-    const detailColor = inTooltip
-      ? "var(--mb-color-tooltip-text-secondary)"
-      : "var(--mb-color-text-secondary)";
-
-    return (
-      <Title order={5} style={{ whiteSpace: "pre", color: detailColor }}>
-        <Separator inTooltip={inTooltip} />
-        {children}
-      </Title>
-    );
-  };
-
-  return (
-    <Tooltip
-      disabled={fullDetailDisplay === fittedDetailDisplay}
-      position="bottom"
-      label={
-        <Flex align="center">
-          <VariationPercent iconSize={TOOLTIP_ICON_SIZE} inTooltip>
-            {display.percentChange}
-          </VariationPercent>
-          <VariationDetails inTooltip>
-            {tooltipFullDetailDisplay}
-          </VariationDetails>
-        </Flex>
-      }
-    >
-      <Flex
-        wrap="wrap"
-        align="center"
-        justify="center"
-        mx="sm"
-        lh="1.2rem"
-        className={cx(
-          DashboardS.fullscreenNormalText,
-          DashboardS.fullscreenNightText,
-          EmbedFrameS.fullscreenNightText,
-        )}
-      >
-        <VariationPercent iconSize={ICON_SIZE}>
-          {fittedChangeDisplay}
-        </VariationPercent>
-        <VariationDetails>{fittedDetailDisplay}</VariationDetails>
-      </Flex>
-    </Tooltip>
   );
 }
